Add AlertManeger and change ExcelData cleaning
Overall, commit is an improvment for UI part and interaction with user.
AlertManager is an addition on top of javafx Alert with convenient
methods to call, e.g. in case of errors stack trace will be
printed in text field in window.

Also this commit changes the way of cleaning temp files, closing
workbooks, etc. Now ExcelData is responsible for that.

Additional:
- In UI button 'New File' changed to 'Back'